drf/udemy/tdd/recipe_api/requirements.txt
Django>=4.0.0,<4.1.0
djangorestframework>=3.13.1,<3.14.0
psycopg2-binary>=2.9.2,<2.10.0
django-extensions>=3.1.5,<3.2.0
